push_the_goddamn_button/requirements.txt
python-fasthtml
uvicorn>=0.29
python-multipart